requirements.txt
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
dataclasses-json==0.6.7
distro==1.9.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
jiter==0.7.1
jsonpatch==1.33
jsonpointer==3.0.0
magic-filter==1.0.12
marshmallow==3.23.1
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.55.1
orjson==3.10.12
packaging==24.2
propcache==0.2.0
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
python-dotenv==1.0.1
python-telegram-bot==21.7
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
tenacity==9.0.0
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
wikipedia==1.4.0
yarl==1.18.0
